requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
qrcode[pil]==7.4.2
Werkzeug==3.0.1